add ResponseEntity to methods of CartController and CartSevice.

+++ springapp/src/main/java/com/examly/springapp/service/CartService.java
@@ -2,6 +2,7 @@ package com.examly.springapp.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.*;
 
 import com.examly.springapp.repository.CartRepository;
 import com.examly.springapp.model.CartModel;
@@ -14,20 +15,40 @@ public class CartService {
     @Autowired
     private CartRepository cartRepository;
 
-    private int numberOfProducts = 0;
-    
-    public void addToCart(String quantity, String id) {
-        // functional requirement last point.
-        if (numberOfProducts == 5) return;
+    /*
+    i. addToCart(String Quantity, String id): This method helps the customer
+    to add the product to the cart.
+    ii. List<CartTempModel> showCart(String id): This method helps to view
+    the cart items.
+    iii. deleteCartItem(String id): This method helps to delete a product from
+    the cart.
+    */
+
+    /* 
+        id is probably product id, quantity as requested by user, 
+        should only add if that much available
+     */
+    public ResponseEntity<String> addToCart(String quantity, String id) {
+        // functional requirement last point needs explaination, not implemented.
+
         // ProductModel product = productEditData(id);
         // int available = Integer.parseInt(product.getQuantity()), asked = Integer.parseInt(quantity);
         // // do not have enough items
-        // if (available < asked) return;
+        // if (available < asked) {
+        //      return ResponseEntity.
+        //      badRequest().
+        //      body(String.format("Cannot add more than %s %s in the cart", product.getQuantity(), product.getName()));
+        // }
         CartModel cartItem = new CartModel();
         // cartItem.setProductName(product.getProductName());
         // cartItem.setQuantity(asked);
         // cartItem.setPrice(Integer.parseInt(product.getPrice()));
         // cartRepository.save(cartItem);
+        // // decrease quantity available
+        // product.setQuantity(new String (available - asked));
+        // productEditSave(product);
+        // return ResponseEntity.ok(String.format("%s %s added to cart", quantity, product.getName()));
+        return ResponseEntity.ok("");
     }
 
     /* 
@@ -36,11 +57,15 @@ public class CartService {
      */
     public List<CartModel> showCart(String id) {
         List<CartModel> cart = cartRepository.findAllByUserIdEmail(id);
-        numberOfProducts = cart.size();
         return cart;
     }
 
-    public void deleteCartItem(String id) {
+    /*
+        id is of CartModel
+     */
+    public ResponseEntity<String> deleteCartItem(String id) {
+        if (!cartRepository.existsById(id)) return ResponseEntity.badRequest().body("Cart Item does not exist.");
         cartRepository.deleteById(id);
+        return ResponseEntity.ok("Cart Item deleted.");
     }
 }
